+++ tests/testteambcast.c
@@ -165,7 +165,7 @@ int main(int argc, char **argv)
     total = TIME() - start;
     
     if(mynode == 0){
-      printf("%d> %ld byte broadcast team all time: %8.3f usec\n",mynode,(long)sz*sizeof(int),
+      printf("%d> %"PRIuPTR" byte broadcast team all time: %8.3f usec\n",mynode,(uintptr_t)sz*sizeof(int),
            ((float)total)/(iters));
       fflush(stdout);
     }
@@ -199,7 +199,7 @@ int main(int argc, char **argv)
     total = TIME() - start;
     
     if(my_col == 0){
-      printf("%d> %ld byte broadcast row team %u time: %8.3f usec\n",mynode,(long)sz*sizeof(int),
+      printf("%d> %"PRIuPTR" byte broadcast row team %u time: %8.3f usec\n",mynode,(uintptr_t)sz*sizeof(int),
            my_row,  ((float)total)/(iters));
       fflush(stdout);
     }
@@ -232,7 +232,7 @@ int main(int argc, char **argv)
     total = TIME() - start;
     
     if(my_row == 0){
-      printf("%d> %ld byte broadcast col team %u time: %8.3f usec\n",mynode,(long)sz*sizeof(int),
+      printf("%d> %"PRIuPTR" byte broadcast col team %u time: %8.3f usec\n",mynode,(uintptr_t)sz*sizeof(int),
            my_col,  ((float)total)/(iters));
       fflush(stdout);
     }
